Cheerleader Details

A dataset containing detailed information about cheerleaders, including personal attributes and social media links.

Usage

cheer_data

Format

A list of cheerleader details where each entry is a list with two components:

table

A tibble with personal information attributes.

links

A character vector of social media profile URLs.

Source

"Extracted from namuwiki cheerleader profile pages."

Details

This dataset is a list where each element corresponds to a cheerleader. Each cheerleader contains:

table

A tibble with 12 rows and 2 columns detailing personal attributes. The columns are:

X1

A character vector describing the type of attribute (e.g., "Birthday", "Birthplace").

X2

A character vector providing the value for each attribute.

links

A character vector containing URLs to the cheerleader's social media profiles.

Examples

data(cheer_data)
# Display details for the first cheerleader
cheer_data[[1]]
#> $table
#> # A tibble: 12 × 2
#>    X1          X2                                                               
#>    <chr>       <chr>                                                            
#>  1 Birthday    "1994 April 22 (30 Years Old)"                                   
#>  2 Birthplace  "Incheon Metropolitan City Bupyeong-Gu"                          
#>  3 Nationality "Korea Region Dependent"                                         
#>  4 Body        "167.7cm, 45.8kg , 235mm, Type A"                                
#>  5 Family      "Parents, 2 Older Sistersdogs Cheongi, Kongi, Gungi"             
#>  6 Nickname    "Sugi, Korea's Top"                                              
#>  7 Agency      "Play With Us Spotainment"                                       
#>  8 Education   "Incheon Bugae Elementary School - Girls' Middle School - Girls'…
#>  9 Debut       "Year 2014 Bucheon Hana Exchange Cheerleader"                    
#> 10 Career      "Bucheon Hana Exchange (2014~2015 / Wkbl) Goyang Orion Orions (2…
#> 11 MBTI        "ESTP"                                                           
#> 12 Link        "<a href=\"https://www.facebook.com/hs0422\" target=\"_blank\"><…
#> 
#> $links
#> [1] "https://www.facebook.com/hs0422"                         
#> [2] "https://www.instagram.com/seo_hsss"                      
#> [3] "https://www.youtube.com/channel/UCrzTDI-wev6gpi0zBzJkqQA"
#> [4] "https://likey.me/seo_hsss"                               
#>