Team Data

A dataset containing information about various cheerleading teams.

Usage

team_data

Format

A data frame with X rows and Y columns:

name

A character vector of team names.

url

A character vector of team namuwiki urls.

color

A character vector of team themesong urls on YouTube.

Details

This dataset includes columns such as team names, locations, and performance metrics.

Examples

data(team_data)
head(team_data)
#>            name
#> 1  Doosan Bears
#> 2 Hanwha Eagles
#> 3    Kia Tigers
#> 4  Kiwoom Heros
#> 5        KT Wiz
#> 6      LG Twins
#>                                                                                                                     url
#> 1          https://en.namu.wiki/w/%EB%91%90%EC%82%B0%20%EB%B2%A0%EC%96%B4%EC%8A%A4/%EC%B9%98%EC%96%B4%EB%A6%AC%EB%8D%94
#> 2          https://en.namu.wiki/w/%ED%95%9C%ED%99%94%20%EC%9D%B4%EA%B8%80%EC%8A%A4/%EC%B9%98%EC%96%B4%EB%A6%AC%EB%8D%94
#> 3          https://en.namu.wiki/w/KIA%20%ED%83%80%EC%9D%B4%EA%B1%B0%EC%A6%88/%EC%B9%98%EC%96%B4%EB%A6%AC%EB%8D%94#s-5.1
#> 4 https://en.namu.wiki/w/%ED%82%A4%EC%9B%80%20%ED%9E%88%EC%96%B4%EB%A1%9C%EC%A6%88/%EC%B9%98%EC%96%B4%EB%A6%AC%EB%8D%94
#> 5                                                  https://en.namu.wiki/w/kt%20wiz/%EC%B9%98%EC%96%B4%EB%A6%AC%EB%8D%94
#> 6                          https://en.namu.wiki/w/LG%20%ED%8A%B8%EC%9C%88%EC%8A%A4/%EC%B9%98%EC%96%B4%EB%A6%AC%EB%8D%94
#>     color                                        song
#> 1 #131230 https://www.youtube.com/watch?v=IeyOr91IuyA
#> 2 #ff6600 https://www.youtube.com/watch?v=45IrJtdGtDk
#> 3 #EC0F32 https://www.youtube.com/watch?v=6RiMyqT3_t0
#> 4 #820023 https://www.youtube.com/watch?v=antR6UYqZKk
#> 5 #000000 https://www.youtube.com/watch?v=bsWBORqEjpI
#> 6 #a50034 https://www.youtube.com/watch?v=CeGiBg9eXG0