Ultra Combo Cheerleader Metrics

A dataset containing metrics for cheerleaders, including social media statistics and team affiliations.

Usage

ultra_combo

Format

A tibble with 138 rows and 15 columns.

Source

"Collected from social media profiles (YouTube, Instagram, TikTok)."

Details

This dataset contains 138 rows and 15 columns with information on cheerleaders' social media metrics and team details. The columns include:

name

A character vector with the names of the cheerleaders.

subs

An integer vector representing the number of subscribers or followers.

views

An integer vector with the number of views.

count

An integer vector indicating the number of items or posts.

cat

A character vector denoting the platform (e.g., "youtube").

likes

An integer vector for the number of likes, with NA indicating missing data.

instagram_followers

An integer vector for Instagram followers, with NA for missing data.

tiktok_followers

An integer vector for TikTok followers, with NA for missing data.

team

A character vector indicating the team the cheerleader belongs to.

link

A character vector containing HTML links to the cheerleader's profile.

team_img

A character vector with filenames of team images.

color

A character vector specifying the team colors in hexadecimal format.

logo

A character vector containing HTML for team logos.

photo

A character vector containing HTML for cheerleader photos.

avg_views_per_video

An integer vector for the average number of views per video, with NA for missing data.

Examples

data(ultra_combo)
head(ultra_combo)
#>            name   subs    views count     cat likes instagram_followers
#> 1   Hyunsuk Seo 125000 55106936   538 youtube    NA                  NA
#> 2 Park Ki-Ryang  56300 13213606   158 youtube    NA                  NA
#> 3     Ha Ji Won   3880        0     0 youtube    NA                  NA
#> 4    Nayeon Kim  11300   894630    93 youtube    NA                  NA
#> 5   Jeong Ga-Ye    211      625     1 youtube    NA                  NA
#> 6    Hannah Kim  87400 27948070    93 youtube    NA                  NA
#>   tiktok_followers          team
#> 1               NA  Doosan Bears
#> 2               NA  Doosan Bears
#> 3               NA Hanwha Eagles
#> 4               NA Hanwha Eagles
#> 5               NA    Kia Tigers
#> 6               NA    Kia Tigers
#>                                                                                                                                   link
#> 1   <a href="https://en.namu.wiki/w/%EC%84%9C%ED%98%84%EC%88%99(%EC%B9%98%EC%96%B4%EB%A6%AC%EB%8D%94)" target="_blank">Hyunsuk Seo</a>
#> 2 <a href="https://en.namu.wiki/w/%EB%B0%95%EA%B8%B0%EB%9F%89(%EC%B9%98%EC%96%B4%EB%A6%AC%EB%8D%94)" target="_blank">Park Ki-Ryang</a>
#> 3                                     <a href="https://en.namu.wiki/w/%ED%95%98%EC%A7%80%EC%9B%90(2002)" target="_blank">Ha Ji Won</a>
#> 4    <a href="https://en.namu.wiki/w/%EA%B9%80%EB%82%98%EC%97%B0(%EC%B9%98%EC%96%B4%EB%A6%AC%EB%8D%94)" target="_blank">Nayeon Kim</a>
#> 5                                         <a href="https://en.namu.wiki/w/%EC%A0%95%EA%B0%80%EC%98%88" target="_blank">Jeong Ga-Ye</a>
#> 6    <a href="https://en.namu.wiki/w/%EA%B9%80%ED%95%9C%EB%82%98(%EC%B9%98%EC%96%B4%EB%A6%AC%EB%8D%94)" target="_blank">Hannah Kim</a>
#>            team_img   color
#> 1  Doosan Bears.png #131230
#> 2  Doosan Bears.png #131230
#> 3 Hanwha Eagles.png #ff6600
#> 4 Hanwha Eagles.png #ff6600
#> 5    Kia Tigers.png #EC0F32
#> 6    Kia Tigers.png #EC0F32
#>                                                                                                             logo
#> 1   <img height=50 src="www/team_logo/Doosan Bears.png"\n  class="team-photo"\n  data-tt="Doosan Bears">\n</img>
#> 2   <img height=50 src="www/team_logo/Doosan Bears.png"\n  class="team-photo"\n  data-tt="Doosan Bears">\n</img>
#> 3 <img height=50 src="www/team_logo/Hanwha Eagles.png"\n  class="team-photo"\n  data-tt="Hanwha Eagles">\n</img>
#> 4 <img height=50 src="www/team_logo/Hanwha Eagles.png"\n  class="team-photo"\n  data-tt="Hanwha Eagles">\n</img>
#> 5       <img height=50 src="www/team_logo/Kia Tigers.png"\n  class="team-photo"\n  data-tt="Kia Tigers">\n</img>
#> 6       <img height=50 src="www/team_logo/Kia Tigers.png"\n  class="team-photo"\n  data-tt="Kia Tigers">\n</img>
#>                                                                                                                                                     photo
#> 1     <img height=50 src="www/cheerleader_img/Hyunsuk Seo.webp"\n class="cheerleader-photo"\n data-team="Doosan Bears"\n data-name="Hyunsuk Seo">\n</img>
#> 2 <img height=50 src="www/cheerleader_img/Park Ki-Ryang.webp"\n class="cheerleader-photo"\n data-team="Doosan Bears"\n data-name="Park Ki-Ryang">\n</img>
#> 3        <img height=50 src="www/cheerleader_img/Ha Ji Won.webp"\n class="cheerleader-photo"\n data-team="Hanwha Eagles"\n data-name="Ha Ji Won">\n</img>
#> 4      <img height=50 src="www/cheerleader_img/Nayeon Kim.webp"\n class="cheerleader-photo"\n data-team="Hanwha Eagles"\n data-name="Nayeon Kim">\n</img>
#> 5       <img height=50 src="www/cheerleader_img/Jeong Ga-Ye.webp"\n class="cheerleader-photo"\n data-team="Kia Tigers"\n data-name="Jeong Ga-Ye">\n</img>
#> 6         <img height=50 src="www/cheerleader_img/Hannah Kim.webp"\n class="cheerleader-photo"\n data-team="Kia Tigers"\n data-name="Hannah Kim">\n</img>
#>   avg_views_per_video age
#> 1              102429  30
#> 2               83630  33
#> 3                  NA  22
#> 4                9619  25
#> 5                 625  28
#> 6              300516  NA